pyproject.toml
[tool.poetry]
name = "election-nlp"
version = "0.1.0"
description = "Reddit election comment NLP analysis pipeline"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
dask = {extras = ["dataframe"], version = "*"}
pandas = "*"
pyarrow = "*"
zstandard = "*"
spacy = "*"
spacy-transformers = "*"
transformers = "*"
torch = "*"
sentence-transformers = "*"
faiss-cpu = "*"
bertopic = "*"
fastapi = "*"
uvicorn = "*"
streamlit = "*"
streamlit-option-menu = "*"
altair = "*"
plotly = "*"
langdetect = "*"
emoji = "*"
vaderSentiment = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
